Add delete perm check

+++ handler/document.go
@@ -99,6 +99,21 @@ func (h *Handler) deleteDocumentHandler(c *gin.Context) {
 		return
 	}
 
+	pfinfo, err := h.db.GetFolderInfo(dinfo.ParentFolderUUID)
+	if err != nil {
+		if err == db.ErrFolderNotFound {
+			c.AbortWithError(http.StatusNotFound, err)
+			return
+		}
+		c.AbortWithError(http.StatusInternalServerError, err)
+		return
+	}
+
+	if !isRelatedUUID(c, pfinfo.OwnerUUID) && pfinfo.Permission != db.FilePermReadWrite {
+		c.AbortWithStatus(http.StatusForbidden)
+		return
+	}
+
 	err = h.db.DeleteDocument(did)
 	if err != nil {
 		c.AbortWithError(http.StatusInternalServerError, err)
